// crescent-ui1/src/app/models/user.model.ts
export class User {
    id: number;
    userId: number;
    userName: string;
    userPassword: string;
    userFirstName: string;
    userLastName: string;
    userEmail: string;
    userRole: string;
    userMobile: string;
    userAddress: string;
    userCity: string;
    userCityPincode: string;
    userAgreementAccepted: boolean;
    userApproved: boolean
}